Major major changes to where things are made.
libaubit4gl.so* is the only one to remain in lib directory.
The libaubit4gl.so will now be known as libaubit4gl-version_build.so...
All others should be made into a directory called plugins-version_build.
This should make it far better for using older binaries with new builds,
as they won't clash.