Add PDF changes from Michael Krauss for EAN-2 Addon and EAN-5 addon.
To quote from his email :
"EAN-5 is a supplement to the EAN-13 and UPC-A barcode used on books. It
is used to give a suggestion for the price of the book.

The EAN-2 is a supplement to the EAN-13 and UPC-A barcodes. It is often
used on magazines and periodicals to indicate an issue number."